Shop
Items can be bought now, but there is no validation or currency withdraw

+++ laravel/database/migrations/2016_04_06_183407_create_inventory_table.php
@@ -12,14 +12,17 @@ class CreateInventoryTable extends Migration
      */
     public function up()
     {
-        Schema::create('inventory', function (Blueprint $table) {
-            $table->integer('user_id');
-            $table->integer('item_id');
-            $table->integer('quantity')->default(0)->unsigned();
-
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('item_id')->references('id')->on('items')->onDelete('cascade');
-        });
+        if(!Schema::hasTable('inventory')){
+        	
+        	Schema::create('inventory', function (Blueprint $table) {
+	            $table->integer('user_id');
+	            $table->integer('item_id');
+	            $table->integer('quantity')->default(0)->unsigned();
+	
+	            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+	            $table->foreign('item_id')->references('id')->on('items')->onDelete('cascade');
+	        });
+        }        		
     }
 
     /**
